pyproject.toml
[project]
name = "molSimplify"
version = "v1.7.3"
license = {file = "LICENSE"}
readme = "README.md"
dependencies = [
    "openbabel-wheel",
    "numpy",
    "scipy",
    "pandas",
    "networkx",
    "scikit-learn",
    "keras",
    "tensorflow",
    "pyyaml",
    "importlib-resources",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-resource-path", "pre-commit"]
docs = [
    "beautifulsoup4",
    "sphinx",
    "readthedocs-sphinx-ext",
    "sphinx_rtd_theme",
]

[project.urls]
Homepage = "https://molsimplify.mit.edu/"
Documentation = "http://hjkgrp.mit.edu/molsimplify-tutorials/"
Repository = "https://github.com/hjkgrp/molSimplify"

[project.scripts]
molsimplify = "molSimplify.__main__:main"
molscontrol = "molSimplify.molscontrol.molscontrol:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"molSimplify.Data" = ["*.dat"]
"molSimplify.Bind" = ["*.dat"]
"molSimplify.Ligands" = ["*.dict", "*.mol", "*.xyz", "*.smi"]
"molSimplify.icons" = ["*.png"]
"molSimplify.python_nn" = ["*.csv"]
"molSimplify.python_krr" = ["*.csv"]
"molSimplify.tf_nn" = ["*/*"]
"molSimplify.molscontrol" = ["*/*"]
